# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Zeiterfassung API
 * OpenAPI spec version: 0.1.0
 */
import type { EmployeeDashboardItem } from './employeeDashboardItem';

export interface DashboardSummary {
  year: number;
  month: number;
  employees: EmployeeDashboardItem[];
}