# lib/api-zod/src/generated/types/employee.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Zeiterfassung API
 * OpenAPI spec version: 0.1.0
 */

export interface Employee {
  id: number;
  firstName: string;
  lastName: string;
  /** @nullable */
  email?: string | null;
  /** Work percentage 0-100 (e.g. 100 = full time, 80 = 80%) */
  pensum: number;
  /** Target annual working hours */
  annualHours: number;
  createdAt: Date;
}