# lib/api-zod/src/generated/types/employeeDashboardItem.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Zeiterfassung API
 * OpenAPI spec version: 0.1.0
 */

export interface EmployeeDashboardItem {
  employeeId: number;
  firstName: string;
  lastName: string;
  pensum: number;
  annualHours: number;
  monthMinutes: number;
  monthTargetMinutes: number;
  yearMinutes: number;
  yearTargetMinutes: number;
}