# lib/api-zod/src/generated/types/employeeMonthSummary.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Zeiterfassung API
 * OpenAPI spec version: 0.1.0
 */

export interface EmployeeMonthSummary {
  employeeId: number;
  firstName: string;
  lastName: string;
  pensum: number;
  annualHours: number;
  /** Total minutes worked this month */
  totalMinutes: number;
  /** Target minutes for this month based on pensum and annualHours */
  targetMinutes: number;
  /** Difference: actual - target (positive = overtime, negative = deficit) */
  differenceMinutes?: number;
  /** Total minutes worked year-to-date */
  yearTotalMinutes: number;
  /** Target minutes year-to-date */
  yearTargetMinutes: number;
}