# lib/api-zod/src/generated/types/employeeMonthlyReport.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Zeiterfassung API
 * OpenAPI spec version: 0.1.0
 */
import type { DailyEntry } from './dailyEntry';
import type { Employee } from './employee';

export interface EmployeeMonthlyReport {
  employee: Employee;
  year: number;
  month: number;
  days: DailyEntry[];
  totalMinutes: number;
  targetMinutes: number;
  differenceMinutes: number;
  yearTotalMinutes: number;
  yearTargetMinutes: number;
}