# lib/api-zod/src/generated/types/employeeUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Zeiterfassung API
 * OpenAPI spec version: 0.1.0
 */

export interface EmployeeUpdate {
  /** @minLength 1 */
  firstName?: string;
  /** @minLength 1 */
  lastName?: string;
  email?: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  pensum?: number;
  /** @minimum 0 */
  annualHours?: number;
}